Refactor webhook.transport.utils.py file
