+++ saleor/plugins/webhook/plugin.py
@@ -109,6 +109,10 @@ from ...webhook.transport.list_stored_payment_methods import (
     get_response_for_stored_payment_method_request_delete,
     invalidate_cache_for_stored_payment_methods,
 )
+from ...webhook.transport.payment import (
+    parse_list_payment_gateways_response,
+    parse_payment_action_response,
+)
 from ...webhook.transport.shipping import (
     get_cache_data_for_shipping_list_methods_for_checkout,
     get_excluded_shipping_data,
@@ -120,17 +124,17 @@ from ...webhook.transport.synchronous.transport import (
     trigger_webhook_sync,
     trigger_webhook_sync_if_not_cached,
 )
-from ...webhook.transport.utils import (
+from ...webhook.transport.taxes import (
     DEFAULT_TAX_CODE,
     DEFAULT_TAX_DESCRIPTION,
+    get_current_tax_app,
+    parse_tax_data,
+)
+from ...webhook.transport.utils import (
     delivery_update,
     from_payment_app_id,
-    get_current_tax_app,
     get_meta_code_key,
     get_meta_description_key,
-    parse_list_payment_gateways_response,
-    parse_payment_action_response,
-    parse_tax_data,
 )
 from ...webhook.utils import get_webhooks_for_event
 from ..base_plugin import BasePlugin, ExcludedShippingMethod
