+++ saleor/plugins/webhook/tests/test_payment_webhook.py
@@ -19,14 +19,16 @@ from ....payment.utils import create_payment_information
 from ....webhook.event_types import WebhookEventAsyncType, WebhookEventSyncType
 from ....webhook.models import Webhook, WebhookEvent
 from ....webhook.transport import signature_for_payload
+from ....webhook.transport.payment import (
+    parse_list_payment_gateways_response,
+    parse_payment_action_response,
+)
 from ....webhook.transport.synchronous.transport import (
     send_webhook_request_sync,
     trigger_webhook_sync,
 )
 from ....webhook.transport.utils import (
     from_payment_app_id,
-    parse_list_payment_gateways_response,
-    parse_payment_action_response,
     to_payment_app_id,
 )
 from .utils import generate_request_headers
