+++ saleor/webhook/tests/test_payment_webhook_utils.py
@@ -4,11 +4,13 @@ from ...core import EventDeliveryStatus, private_storage
 from ...core.models import EventDelivery, EventPayload
 from ...payment import TransactionKind
 from ..const import APP_ID_PREFIX
+from ..transport.payment import (
+    parse_list_payment_gateways_response,
+    parse_payment_action_response,
+)
 from ..transport.utils import (
     clear_successful_delivery,
     from_payment_app_id,
-    parse_list_payment_gateways_response,
-    parse_payment_action_response,
     to_payment_app_id,
 )
